Rename "index" files to descriptive names
