+++ p2panda-js/src/identity/index.ts
@@ -1,0 +1,1 @@
+export * from './keyPair';
