+++ p2panda-js/src/index.ts
@@ -9,8 +9,8 @@ if (!globalThis.fetch) {
   globalThis.Headers = Headers;
 }
 
-export { createKeyPair, recoverKeyPair } from './keyPair';
-export { default as Session } from './session';
+export { createKeyPair, recoverKeyPair } from './identity';
+export { Session } from './session';
 export { default as wasm } from './wasm';
 
 export * from './types';
