+++ p2panda-js/src/message/index.ts
@@ -1,33 +1,1 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-import debug from 'debug';
-
-import { Session } from '~/index';
-import { FieldsTagged } from '~/types';
-
-import { KeyPair, MessageFields } from 'wasm-web';
-
-export type Context = {
-  keyPair: KeyPair;
-  schema: string;
-  session: Session;
-};
-
-const log = debug('p2panda-js:message');
-
-/**
- * Returns a message fields instance for the given field contents and schema.
- */
-export const getMessageFields = async (
-  session: Session,
-  fields: FieldsTagged,
-): Promise<MessageFields> => {
-  const { MessageFields } = await session.loadWasm();
-
-  const messageFields = new MessageFields();
-  for (const k of Object.keys(fields)) {
-    messageFields.add(k, fields[k]['type'], fields[k]['value']);
-  }
-  log('getMessageFields', messageFields.toString());
-  return messageFields;
-};
+export * from './message';
